♻️ Replace Homepage Presenter with decorator
